mlb_bot/requirements.txt
python-telegram-bot==20.7
pybaseball==2.2.5
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
scipy==1.12.0
requests==2.31.0
python-dotenv==1.0.0
mlb-statsapi==1.6.1
aiohttp==3.9.3
httpx==0.26.0
tabulate==0.9.0
google-cloud-bigquery==3.14.1
db-dtypes==1.2.0
